refactor(code-preview): split syntect backend into bundle render shell and semantics modules
